intel-sound: Synchronize with 20230625's WHENCE file
The Intel's sound firmware get new version in a new file and the general
Link then gets updated to point to the new version/file.
So the old versions/files become useless cruft, but for some reason they
do not get removed from the WHENCE file.

Don't add the old files to ``debian/copyright``'s Files-Excluded
section yet as that would violate Policy 3.2.2. They can be added with a
new upstream version.

Now that all the ``dsp_fw_*`` files are symlinks into the ``avs``
directory, update ``debian/copyright`` to point to the actual files.

The intel-sound firmware is now synchronized with 20230625's WHENCE
file, but unlike with other firmware packages, NOT all files are present
in the defines file.