d/README.source: Update description wrt upstream
Previously it mentioned that upstream did not make tagged releases, but
that has changed (since 20190312), so update the description to reflect
that.

Also change the upstream URL from git to https protocol.
The https variant is easier to use for most people, it's used in
``debian/watch`` and via that URL one can find the git variant too.

Also remove "``" which is used in Markdown to indicate code, but this is
not Markdown.