misc-nonfree: Move Nvidia graphics firmware to new firmware-nvidia-graphics package
The misc-nonfree package is for firmware files 'which are not
individually large enough to warrant a standalone package'.

This is not applicable to the Nvidia graphics firmware files anymore,
so move the Nvidia graphics firmware files into the nvidia-graphics
package.

Add ``firmware-misc-nonfree.NEWS`` file so the user get informed that
they should now install the new package to get the Nvidia graphics
firmware.