misc-nonfree: Move Intel graphics firmware to new firmware-intel-graphics package
The misc-nonfree package is for firmware files 'which are not
individually large enough to warrant a standalone package'.

This is not applicable to the Intel graphics (mostly i915) firmware
files anymore, so move them into their own package.

Update ``firmware-misc-nonfree.NEWS`` file so the user get informed that
they should now install the new package to get the Intel graphics
firmware.