misc-nonfree: Move MediaTek/Ralink firmware to new firmware-mediatek package
The misc-nonfree package is for firmware files 'which are not
individually large enough to warrant a standalone package'.

This is not applicable to the MediaTek and Ralink (a MediaTek company)
firmware files anymore, so move them into their own package.

As the lintian-overrides from the misc-nonfree package were for the
``scp.img`` files, add those overrides to this new package and remove
the lintian-overrides file from firmware-misc-nonfree.

Update ``firmware-misc-nonfree.NEWS`` file so the user get informed that
they should now install the new package to get the MediaTek firmware.