ci: extract-source: Install `dpkg-dev` early and explicitly
The next statement runs ``debian/rules`` to determine which version of
the ``linux-support`` package to install. Executing ``debian/rules``
makes a call to ``dpkg-parsechangelog`` which is part of ``dpkg-dev``
package. As it wasn't installed yet, that resulted in these errors:

```
sh: 1: dpkg-parsechangelog: not found
sh: 1: dpkg-parsechangelog: not found
```

Non-fatal errors, but still errors, so install ``dpkg-dev`` early and
explicitly.