d/copyright: libertas: Update copyright data
Various files in the ``mrvl`` directory are licensed under
``LICENCE.NXP``, but there was already a ``binary-redist-NXP``
identifier, which contained the text from ``LICENSE.sdma_firmware``.
As this is also by NXP, rename the identifier to
``binary-redist-NXP-SDMA``.

This made room for ``binary-redist-NXP-1`` to have the text of
``LICENCE.NXP``. The '-1' suffix is because there's also a
``LICENSE.nxp`` license file which is used for firmware files in the
``nxp`` directory.

Also (slightly) rename the ``Marvell`` license identifier to closer
match their LICENCE file name in the upstream repo.