d/copyright: Update license for Intel qat* firmware
On 2023-03-08, the ``LICENCE.qat_firmware`` was changed/updated.

As binary-redist-Intel and binary-redist-Intel-1 had the same contents,
'repurpose' binary-redist-Intel-1 to have the license text from the
new/updated ``LICENCE.qat_firmware``.
Also add a note in Comment that binary-redist-Intel-1 and
binary-redist-Intel-2 are the same except for 1 word.

Also change the license ID for ``intel/catpt/*`` to binary-redist-Intel
as its contents is the same as (f.e.) the i915 license and add a comment
to binary-redist-Intel to list all the upstream license files whose text
is the same, except for the copyright statements.