d/copyright: misc-nonfree: Add Silicon Labs copyright data
It's not really clear (to me) if ``wfx/brd*.pds`` files also fall under
this license as they're generated from ``*.pds.in`` files coming from a
different (git/GitHub) repository which doesn't seem to have a license
file. Detail in ``WHENCE``.

OTOH, the ``WHENCE`` file gives the impression that they are licensed
under the same license as the firmware.

The firmware-free repo neither excludes it (due to being unfree), but
it's also not included (in ``debian/config/*``).

While at it, update the firmware's version number.