d/b/{gencontrol,check_upstream}.py: Support wildcards in file lists
Add support for wildcards in the files list, so it's not necessary to
list every single file that's added.  Also add a files-excluded field,
to allow excluding unwanted files that match a broad pattern.

This is more complicated than it ought to be because:

- pathlib.Path.glob() currently uses the wrong stat() variant and does
  not list the dangling symlinks that are present in some per-package
  config directories.  We already do a directory walk to find all
  files and report unused ones, so match against that file list rather
  than globbing.

- pathlib.Path.glob() supports the '**' wildcard but the match()
  method does not.  Implement our own translation of patterns to
  regexps instead of using match().

Update the generation of FILES and LINKS variables in rules.gen to
work for filenames that have spaces in them, which could not be
specified in file lists but can now be matched by wildcards.

Thanks: James Addison <jay@jp-hosting.net>
Closes: #1035505