misc-nonfree: Synchronize with 20230625's WHENCE file
Add and update defines entries and file stanzas based on the WHENCE file
and commit messages.

As this is the package for 'the rest of the files', don't claim that all
`File:` and `Link:` items from 20230625's WHENCE file are now present in
this defines file, as they are not.