Fix the "git describe" issue when triggered on annotated tag
When we push a signed and annotated tag to GitHub, this action
will be triggered. Using the default checkout action will
only provide a checkout where HEAD is on an unannotated tag,
even if you add `fetch-depth: 0` and `fetch-tags: true`.
That means that `git describe`, without --tags, will fail
to find the current tag being built, or it will show the
previous annotated tag.

If you run `git describe --tags`, the current tag will be
shown correctly, but I don't want to modify the forked
project's build script for an issue that appears to be with
GitHub's checkout action.

Forcing the checkout ref to be github.ref, seems to resolve
the aforementioned issue for tags, but could be racey, so lets
limit the usage of this to only building for tags.

Checkout https://github.com/actions/checkout/issues/649 for more
context.